Alterando pop-up de alert para biblioteca SweetAlert

+++ assets/js/publicacoes.js
@@ -19,7 +19,7 @@ function criarPublicacao(evento) {
     }).done(function () {
         window.location = "/home";
     }).fail(function () {
-        alert("Erro ao criar publicação.")
+        Swal.fire("Ops...", "Erro ao criar publicação", "error");
     })
 }
 
@@ -45,7 +45,7 @@ function curtirPublicacao(evento) {
         elementoClicado.removeClass(('curtir-publicacao'));
 
     }).fail(function () {
-        alert("Erro ao curtir publicação.")
+        Swal.fire("Ops...", "Erro ao curtir publicação", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     });
@@ -73,7 +73,7 @@ function descurtirPublicacao(evento) {
         elementoClicado.addClass('curtir-publicacao');
 
     }).fail(function () {
-        alert("Erro ao curtir publicação.")
+        Swal.fire("Ops...", "Erro ao descurtir publicação", "error");
     }).always(function () {
         elementoClicado.prop('disabled', false);
     });
@@ -92,9 +92,12 @@ function atualizarPublicacao() {
             conteudo: $('#conteudo').val()
         }
     }).done(function () {
-        alert("Publicação editada com sucesso!");
+        Swal.fire('Sucesso!', 'Publicação atualizada', 'success')
+            .then(function () {
+                window.location = "/home";
+            })
     }).fail(function () {
-        alert("Erro ao editar publicação.");
+        Swal.fire("Ops...", "Erro ao editar publicação", "error");
     }).always(function () {
         $('#atualizar-publicacao').prop('disabled', false);
     })
@@ -103,20 +106,30 @@ function atualizarPublicacao() {
 function deletarPublicacao(evento) {
     evento.preventDefault();
 
-    const elementoClicado = $(evento.target);
-    const publicacao = elementoClicado.closest('div');
-    const publicacaoId = publicacao.data('publicacao-id');
+    Swal.fire({
+        title: "Atenção!",
+        text: "Tem certeza que deseja excluir esta publicação?",
+        showCancelButton: true,
+        cancelButtonText: "Cancelar",
+        icon: "warning"
+    }).then(function (confirmacao) {
+        if (!confirmacao.value) return;
 
-    elementoClicado.prop('disabled', true);
+        const elementoClicado = $(evento.target);
+        const publicacao = elementoClicado.closest('div');
+        const publicacaoId = publicacao.data('publicacao-id');
 
-    $.ajax({
-        url: `/publicacoes/${publicacaoId}`,
-        method: "DELETE",
-    }).done(function () {
-        publicacao.fadeOut("slow", function() {
-            $(this).remove();
+        elementoClicado.prop('disabled', true);
+
+        $.ajax({
+            url: `/publicacoes/${publicacaoId}`,
+            method: "DELETE",
+        }).done(function () {
+            publicacao.fadeOut("slow", function () {
+                $(this).remove();
+            });
+        }).fail(function () {
+            Swal.fire("Ops...", "Erro ao excluir publicação", "error");
         });
-    }).fail(function () {
-        alert("Erro ao excluir publicação")
-    });
+    })
 }
